requirements.txt
requests
python-dotenv
opencv-python
ultralytics
torch
numpy
Flask
